+++ Monofoxe/Monofoxe.Engine/SceneSystem/SceneMgr.cs
@@ -27,7 +27,11 @@ namespace Monofoxe.Engine.SceneSystem
 		/// </summary>
 		public static Scene CurrentScene {get; private set;}
 
-		
+		/// <summary>
+		/// True, when the current frame is the frame on which FixedUpdate will be executed.
+		/// </summary>
+		public static bool IsFixedUpdateFrame {get; private set;}
+
 		/// <summary>
 		/// Counts time until next fixed update.
 		/// </summary>
@@ -199,7 +203,6 @@ namespace Monofoxe.Engine.SceneSystem
 		/// </summary>
 		public static event SceneMgrEventDelegate OnPostDrawGUI;
 
-
 		/// <summary>
 		/// Executes Fixed Update events.
 		/// </summary>
@@ -207,8 +210,10 @@ namespace Monofoxe.Engine.SceneSystem
 		{
 			_fixedUpdateTimer += gameTime.ElapsedGameTime.TotalSeconds;
 
+			IsFixedUpdateFrame = false;
 			if (_fixedUpdateTimer >= GameMgr.FixedUpdateRate)
 			{
+				IsFixedUpdateFrame = true;
 				var overflow = (int)(_fixedUpdateTimer / GameMgr.FixedUpdateRate); // In case of lags.
 				_fixedUpdateTimer -= GameMgr.FixedUpdateRate * overflow;